Agent takes on langchainagent and cleanup across
